# docs/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Cheesy Arena API
  version: 1.0.0-draft
  description: >-
    Versioned API introduced alongside the existing server-rendered interface.
    This document is expanded as each migration phase is implemented.
servers:
  - url: /api/v1
paths:
...
  /bracket:
    get:
      operationId: getBracket
      summary: Get the playoff bracket as structured data
      parameters:
        - in: query
          name: activeMatch
          schema:
            type: string
            enum: [current, saved]
      responses:
        "200":
          description: Playoff bracket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BracketResponse"
  /event:
    get:
      operationId: getPublicEvent
      summary: Get public event configuration
      responses:
        "200":
          description: Public event configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicEventResponse"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /matches/{type}:
    get:
      operationId: listMatches
      summary: List matches and their latest results
      parameters:
        - in: path
          name: type
          required: true
          schema:
            $ref: "#/components/schemas/MatchType"
      responses:
        "200":
          description: Matches ordered by type order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchListResponse"
        "400":
          description: Invalid match type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /matches/{matchId}/stations/{stationId}/logs:
    get:
      operationId: getMatchStationLogs
      summary: Get bounded and optionally sampled driver station log rows
      parameters:
        - in: path
          name: matchId
          required: true
          schema:
            type: integer
        - in: path
          name: stationId
          required: true
          schema:
            type: string
            enum: [R1, R2, R3, B1, B2, B3]
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 10000
            default: 1000
        - in: query
          name: sampleEvery
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 1
      responses:
        "200":
          description: Match log details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchLogDetailResponse"
        "400":
          description: Invalid query parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Match or station log not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /teams/{teamId}/avatar:
    get:
      operationId: getTeamAvatar
      summary: Get a team avatar or the default avatar
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Avatar image
          content:
            image/png:
              schema:
                type: string
                contentEncoding: binary
  /admin/awards:
    get:
      operationId: listAdminAwards
      summary: List awards for administration
      security: [{sessionCookie: []}]
      responses:
        "200": {$ref: "#/components/responses/AdminAwardList"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
    post:
      operationId: createAdminAward
      summary: Create a judged award
      security: [{sessionCookie: []}]
      parameters: [{$ref: "#/components/parameters/CsrfToken"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AdminAwardInput"}
      responses:
        "201": {$ref: "#/components/responses/AdminAwardItem"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /admin/awards/{id}:
    parameters: [{$ref: "#/components/parameters/ResourceId"}]
    patch:
      operationId: updateAdminAward
      summary: Update a judged award
      security: [{sessionCookie: []}]
      parameters: [{$ref: "#/components/parameters/CsrfToken"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AdminAwardInput"}
      responses:
        "200": {$ref: "#/components/responses/AdminAwardItem"}
        "404": {$ref: "#/components/responses/NotFound"}
        "422": {$ref: "#/components/responses/ValidationError"}
    delete:
      operationId: deleteAdminAward
      summary: Delete a judged award
      security: [{sessionCookie: []}]
      parameters: [{$ref: "#/components/parameters/CsrfToken"}]
      responses:
        "204": {description: Award deleted}
        "404": {$ref: "#/components/responses/NotFound"}
  /admin/scheduled-breaks:
    get:
      operationId: listAdminScheduledBreaks
      summary: List scheduled breaks for administration
      security: [{sessionCookie: []}]
      responses:
        "200":
          description: Scheduled breaks
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AdminScheduledBreakListResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /admin/scheduled-breaks/{id}:
    parameters: [{$ref: "#/components/parameters/ResourceId"}]
    patch:
      operationId: updateAdminScheduledBreak
      summary: Update the description of a scheduled break
      security: [{sessionCookie: []}]
      parameters: [{$ref: "#/components/parameters/CsrfToken"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AdminScheduledBreakInput"}
      responses:
        "200":
          description: Updated scheduled break
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AdminScheduledBreakResponse"}
        "404": {$ref: "#/components/responses/NotFound"}
  /admin/sponsor-slides:
    get:
      operationId: listAdminSponsorSlides
      summary: List sponsor slides for administration
      security: [{sessionCookie: []}]
      responses:
        "200":
          description: Sponsor slides
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SponsorSlideListResponse"}
    post:
      operationId: createAdminSponsorSlide
      summary: Create a sponsor slide
      security: [{sessionCookie: []}]
      parameters: [{$ref: "#/components/parameters/CsrfToken"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AdminSponsorSlideInput"}
      responses:
        "201":
          description: Created sponsor slide
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SponsorSlideResponse"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /admin/sponsor-slides/{id}:
    parameters: [{$ref: "#/components/parameters/ResourceId"}]
    patch:
      operationId: updateAdminSponsorSlide
      summary: Update a sponsor slide
      security: [{sessionCookie: []}]
      parameters: [{$ref: "#/components/parameters/CsrfToken"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AdminSponsorSlideInput"}
      responses:
        "200":
          description: Updated sponsor slide
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SponsorSlideResponse"}
        "404": {$ref: "#/components/responses/NotFound"}
        "422": {$ref: "#/components/responses/ValidationError"}
    delete:
      operationId: deleteAdminSponsorSlide
      summary: Delete a sponsor slide
      security: [{sessionCookie: []}]
      parameters: [{$ref: "#/components/parameters/CsrfToken"}]
      responses:
        "204": {description: Sponsor slide deleted}
        "404": {$ref: "#/components/responses/NotFound"}
  /admin/sponsor-slides/reorder:
    post:
      operationId: reorderAdminSponsorSlides
      summary: Replace the display order of all sponsor slides
      security: [{sessionCookie: []}]
      parameters: [{$ref: "#/components/parameters/CsrfToken"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AdminReorderInput"}
      responses:
        "200":
          description: Reordered sponsor slides
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SponsorSlideListResponse"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /admin/lower-thirds:
    get:
      operationId: listAdminLowerThirds
      summary: List lower thirds for administration
      security: [{sessionCookie: []}]
      responses:
        "200":
          description: Lower thirds
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AdminLowerThirdListResponse"}
    post:
      operationId: createAdminLowerThird
      summary: Create a lower third
      security: [{sessionCookie: []}]
      parameters: [{$ref: "#/components/parameters/CsrfToken"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AdminLowerThirdInput"}
      responses:
        "201": {$ref: "#/components/responses/AdminLowerThirdItem"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /admin/lower-thirds/{id}:
    parameters: [{$ref: "#/components/parameters/ResourceId"}]
    patch:
      operationId: updateAdminLowerThird
      summary: Update a lower third
      security: [{sessionCookie: []}]
      parameters: [{$ref: "#/components/parameters/CsrfToken"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AdminLowerThirdInput"}
      responses:
        "200": {$ref: "#/components/responses/AdminLowerThirdItem"}
        "404": {$ref: "#/components/responses/NotFound"}
        "422": {$ref: "#/components/responses/ValidationError"}
    delete:
      operationId: deleteAdminLowerThird
      summary: Delete a lower third
      security: [{sessionCookie: []}]
      parameters: [{$ref: "#/components/parameters/CsrfToken"}]
      responses:
        "204": {description: Lower third deleted}
        "404": {$ref: "#/components/responses/NotFound"}
  /admin/lower-thirds/reorder:
    post:
      operationId: reorderAdminLowerThirds
      summary: Replace the display order of all lower thirds
      security: [{sessionCookie: []}]
      parameters: [{$ref: "#/components/parameters/CsrfToken"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AdminReorderInput"}
      responses:
        "200":
          description: Reordered lower thirds
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AdminLowerThirdListResponse"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /admin/teams:
    get:
      operationId: listAdminTeams
      summary: List complete administrative team records with secrets masked
      security: [{sessionCookie: []}]
      responses:
        "200":
          description: Teams and team-list state
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AdminTeamListResponse"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
    post:
      operationId: importAdminTeams
      summary: Start a bulk team import job
      security: [{sessionCookie: []}]
      parameters: [{$ref: "#/components/parameters/CsrfToken"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AdminTeamImportInput"}
      responses:
        "202": {$ref: "#/components/responses/AdminJobItem"}
        "409": {$ref: "#/components/responses/Conflict"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /admin/teams/{id}:
    parameters: [{$ref: "#/components/parameters/ResourceId"}]
    get:
      operationId: getAdminTeam
      summary: Get an administrative team record with its WPA key masked
      security: [{sessionCookie: []}]
      responses:
        "200": {$ref: "#/components/responses/AdminTeamItem"}
        "404": {$ref: "#/components/responses/NotFound"}
    patch:
      operationId: updateAdminTeam
      summary: Update team metadata with explicit WPA key secret semantics
      security: [{sessionCookie: []}]
      parameters: [{$ref: "#/components/parameters/CsrfToken"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AdminTeamUpdateInput"}
      responses:
        "200": {$ref: "#/components/responses/AdminTeamItem"}
        "404": {$ref: "#/components/responses/NotFound"}
        "422": {$ref: "#/components/responses/ValidationError"}
    delete:
      operationId: deleteAdminTeam
      summary: Delete a team before qualification matches exist
      security: [{sessionCookie: []}]
      parameters: [{$ref: "#/components/parameters/CsrfToken"}]
      responses:
        "204": {description: Team deleted}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /admin/teams/refresh:
    post:
      operationId: refreshAdminTeams
      summary: Start a job to refresh every team from The Blue Alliance
      security: [{sessionCookie: []}]
      parameters: [{$ref: "#/components/parameters/CsrfToken"}]
      responses:
        "202": {$ref: "#/components/responses/AdminJobItem"}
  /admin/teams/wpa-keys:
    post:
      operationId: generateAdminTeamWpaKeys
      summary: Generate missing or replacement WPA keys without returning them
      security: [{sessionCookie: []}]
      parameters: [{$ref: "#/components/parameters/CsrfToken"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AdminTeamWpaKeysInput"}
      responses:
        "200":
          description: Number of teams updated
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AdminTeamWpaKeysResponse"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /admin/jobs/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: {type: string, format: uuid}
    get:
      operationId: getAdminJob
      summary: Get background job progress and outcome
      security: [{sessionCookie: []}]
      responses:
        "200": {$ref: "#/components/responses/AdminJobItem"}
        "404": {$ref: "#/components/responses/NotFound"}
  /admin/judging-schedule:
    get:
      operationId: getAdminJudgingSchedule
      summary: Get judging generation parameters and slots ordered by judge and time
      security: [{sessionCookie: []}]
      responses:
        "200": {$ref: "#/components/responses/AdminJudgingScheduleItem"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
    post:
      operationId: generateAdminJudgingSchedule
      summary: Generate and atomically retain a judging schedule
      security: [{sessionCookie: []}]
      parameters: [{$ref: "#/components/parameters/CsrfToken"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/JudgingScheduleParams"}
      responses:
        "201": {$ref: "#/components/responses/AdminJudgingScheduleItem"}
        "409": {$ref: "#/components/responses/Conflict"}
        "422": {$ref: "#/components/responses/ValidationError"}
    delete:
      operationId: clearAdminJudgingSchedule
      summary: Clear all judging slots
      security: [{sessionCookie: []}]
      parameters: [{$ref: "#/components/parameters/CsrfToken"}]
      responses:
        "204": {description: Judging schedule cleared}
  /admin/settings/{section}:
    parameters:
      - in: path
        name: section
        required: true
        schema: {type: string, enum: [event, integrations, network, hardware, game]}
    get:
      operationId: getAdminSettingsSection
      summary: Get one settings section with secrets represented only as configured flags
      security: [{sessionCookie: []}]
      responses:
        "200":
          description: Settings section
          content:
            application/json:
              schema:
                oneOf:
                  - {$ref: "#/components/schemas/EventSettingsResponse"}
                  - {$ref: "#/components/schemas/IntegrationSettingsResponse"}
                  - {$ref: "#/components/schemas/NetworkSettingsResponse"}
                  - {$ref: "#/components/schemas/HardwareSettingsResponse"}
                  - {$ref: "#/components/schemas/GameSettingsResponse"}
        "404": {$ref: "#/components/responses/NotFound"}
    patch:
      operationId: updateAdminSettingsSection
      summary: Partially update one settings section
      security: [{sessionCookie: []}]
      parameters: [{$ref: "#/components/parameters/CsrfToken"}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - {$ref: "#/components/schemas/EventSettingsInput"}
                - {$ref: "#/components/schemas/IntegrationSettingsInput"}
                - {$ref: "#/components/schemas/NetworkSettingsInput"}
                - {$ref: "#/components/schemas/HardwareSettingsInput"}
                - {$ref: "#/components/schemas/GameSettingsInput"}
      responses:
        "200":
          description: Updated settings section
          content:
            application/json:
              schema:
                oneOf:
                  - {$ref: "#/components/schemas/EventSettingsResponse"}
                  - {$ref: "#/components/schemas/IntegrationSettingsResponse"}
                  - {$ref: "#/components/schemas/NetworkSettingsResponse"}
                  - {$ref: "#/components/schemas/HardwareSettingsResponse"}
                  - {$ref: "#/components/schemas/GameSettingsResponse"}
        "409": {$ref: "#/components/responses/Conflict"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /admin/database/backups:
    post:
      operationId: createAdminDatabaseBackup
      summary: Create a server-side snapshot of the event database
      security: [{sessionCookie: []}]
      parameters:
        - {$ref: "#/components/parameters/CsrfToken"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      responses:
        "201":
          description: Backup created
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DatabaseBackupResponse"}
        "409": {$ref: "#/components/responses/Conflict"}
...
  /admin/publishing/{resource}:
    parameters:
      - in: path
        name: resource
        required: true
        schema: {type: string, enum: [teams, matches, rankings, alliances, awards]}
    post:
      operationId: publishAdminResource
      summary: Start an idempotent background TBA publishing job
      security: [{sessionCookie: []}]
      parameters:
        - {$ref: "#/components/parameters/CsrfToken"}
        - {$ref: "#/components/parameters/IdempotencyKey"}
      responses:
        "202":
          description: Publishing job accepted
          content:
            application/json:
              schema: {$ref: "#/components/schemas/PublishingResponse"}
        "409": {$ref: "#/components/responses/Conflict"}
        "422": {$ref: "#/components/responses/ValidationError"}
...
  /admin/referee/bootstrap:
    get:
      operationId: getRefereePanelBootstrap
      summary: Get the referee panel read-model and stream URL
      security: [{sessionCookie: []}]
      responses:
        "200": {description: Current match, clock, fouls/cards, station bypass, and scoring readiness}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_token
...
  responses:
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    MatchType:
      type: string
      enum: [test, practice, qualification, playoff]
...
    PublicEventResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          $ref: "#/components/schemas/PublicEvent"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    MatchListResponse:
      type: object
      required: [data, meta]
      properties:
        data: {type: array, items: {$ref: "#/components/schemas/MatchWithResult"}}
        meta: {$ref: "#/components/schemas/ResponseMeta"}
...
    BracketResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [bracketType, activeMatchId, matchups]
          properties:
            bracketType: {type: string, enum: [double, double4, "16", "8", "4", "2"]}
            activeMatchId: {type: [integer, "null"]}
            matchups: {type: array, items: {type: object}}
        meta: {$ref: "#/components/schemas/ResponseMeta"}
...